+++ src/main/java/com/modugarden/common/error/exception/custom/BusinessException.java
@@ -1,0 +1,14 @@
+package com.modugarden.common.error.exception.custom;
+
+import com.modugarden.common.error.enums.ErrorMessage;
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException{
+    private final ErrorMessage errorMessage;
+
+    public BusinessException(ErrorMessage errorMessage) {
+        super(errorMessage.getMessage());
+        this.errorMessage = errorMessage;
+    }
+}
