+++ src/main/java/com/modugarden/common/error/exception/custom/LoginCancelException.java
@@ -1,0 +1,10 @@
+package com.modugarden.common.error.exception.custom;
+
+import com.modugarden.common.error.enums.ErrorMessage;
+
+public class LoginCancelException extends BusinessException{
+
+    public LoginCancelException(ErrorMessage errorMessage) {
+        super(errorMessage);
+    }
+}
